04_Production_RAG/pyproject.toml
[project]
name = "04-producton-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grandalf>=0.8",
    "jupyter>=1.1.1",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-experimental>=0.3.0",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.5.0",
    "langgraph-cli[inmem]>=0.4.2",
    "langsmith>=0.4.4",
    "nltk>=3.8.1",
    "pandas>=2.3.0",
    "pymupdf>=1.26.1",
]
